Handle broken pipe in session list JSON output
